tts_vertex.py
from google.cloud import aiplatform_v1beta1
from google.cloud import storage
import uuid

def synthesize_text(text: str, bucket_name: str) -> str:
    client = aiplatform_v1beta1.PredictionServiceClient()
    endpoint = "projects/YOUR_PROJECT_ID/locations/us-central1/publishers/google/models/text-bison"

    # Example TTS prompt
    payload = {
        "instances": [{"prompt": f"Synthesize speech for: {text}"}],
        "parameters": {"temperature": 0.7}
    }

    response = client.predict(endpoint=endpoint, instances=payload["instances"], parameters=payload["parameters"])
    audio_content = response.predictions[0].get("audio", "")

    if not audio_content:
        raise Exception("Vertex AI did not return audio.")

    # Save audio to GCS
    filename = f"{uuid.uuid4()}.mp3"
    blob_path = f"audio/{filename}"
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(blob_path)
    blob.upload_from_string(audio_content, content_type="audio/mpeg")

    return f"https://storage.googleapis.com/{bucket_name}/{blob_path}"