requirements.txt
Initial upload for Cloud Run deployment
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.49.0
google-cloud-storage==2.14.0